requirements.txt
asgiref==3.7.2
Babel==2.13.0
backports.zoneinfo==0.2.1
dj-database-url==2.1.0
Django==3.2.22
django-cors-headers==4.1.0
django-extra-views==0.14.0
django-haystack==3.2.1
django-oscar==3.2
django-oscar-api==3.2.4
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.13.0
gunicorn==21.2.0
importlib-metadata==6.7.0
packaging==23.2
phonenumbers==8.13.22
Pillow==9.5.0
psycopg==3.1.12
psycopg2-binary==2.9.9
purl==1.6
pycountry==22.3.5
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
zipp==3.15.0
